En proceso

+++ ChallengeDiWork/Modelo/Desperfecto.cs
@@ -9,7 +9,7 @@ namespace ChallengeDiWork.Logica
     public class Desperfecto
     {
         public string Descripcion { get; set; }
-        public int ManoDeObra { get; set; }
+        public decimal ManoDeObra { get; set; }
         public decimal Tiempo { get; set; }
         public Vehiculo Vehiculo { get; set; }
         public List<Repuesto> Repuestos { get; set; }
